refactor: organize anomaly algorithms into packages
