+++ anomavision/algorithm/__init__.py
@@ -1,0 +1,1 @@
+"""Anomaly detection algorithm implementations."""
