+++ anomavision/algorithm/padim/__init__.py
@@ -1,0 +1,6 @@
+"""PaDiM anomaly detection algorithms."""
+
+from .padim import Padim
+from .padim_lite import PadimLite, load_padim_lite
+
+__all__ = ["Padim", "PadimLite", "load_padim_lite"]
