+++ anomavision/algorithm/patchcore/__init__.py
@@ -1,0 +1,5 @@
+"""Lightweight PatchCore anomaly detection algorithms."""
+
+from .patchcore import PatchCore, build_patchcore_from_stats
+
+__all__ = ["PatchCore", "build_patchcore_from_stats"]
